Files uploaded
Stock ticking application

UI : using C#, WPF-MVVM ( StocksDashboard.zip)
The MainWindow has been designed in a simple way to display two stocks (stock 1 and Stock 2) showing their names and corresponding prices ticking every second. The user can click on any of the stock prices to check the history of prices which is displayed on a simple popup in a grid format displaying the date/time and the price at that time. The popup can be closed just by clicking anywhere outside the popup. The stock price block is displayed in red indicating the price is lower than its previous value or in green indicating the price is higher than its previous value or in white if current and previous prices are equal. Binding to the UI elements is done through the view model. 

Service: WCF ( Ticker1Service.zip)
A WCF service has been created with two different operations to provide prices for stock 1 and stock 2 separately. Once the service is running, the UI application consumes the service to fetch the current prices for both the stocks every second. The logic in the service randomly gives a value with 2 decimal points.